Move notebook chat code to a separate folder
