+++ src/vs/workbench/contrib/chat/browser/chatEditing/notebook/chatEditingModifiedNotebookDiff.ts
@@ -3,13 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { DisposableStore } from '../../../../../base/common/lifecycle.js';
-import { computeDiff } from '../../../notebook/common/notebookDiff.js';
-import { INotebookEditorModelResolverService } from '../../../notebook/common/notebookEditorModelResolverService.js';
-import { INotebookLoggingService } from '../../../notebook/common/notebookLoggingService.js';
-import { INotebookEditorWorkerService } from '../../../notebook/common/services/notebookWorkerService.js';
-import { IEditSessionEntryDiff } from '../../common/chatEditingService.js';
-import { ISnapshotEntry } from './chatEditingModifiedFileEntry.js';
+import { DisposableStore } from '../../../../../../base/common/lifecycle.js';
+import { computeDiff } from '../../../../notebook/common/notebookDiff.js';
+import { INotebookEditorModelResolverService } from '../../../../notebook/common/notebookEditorModelResolverService.js';
+import { INotebookLoggingService } from '../../../../notebook/common/notebookLoggingService.js';
+import { INotebookEditorWorkerService } from '../../../../notebook/common/services/notebookWorkerService.js';
+import { IEditSessionEntryDiff } from '../../../common/chatEditingService.js';
+import { ISnapshotEntry } from '../chatEditingModifiedFileEntry.js';
 
 export class ChatEditingModifiedNotebookDiff {
 	static NewModelCounter: number = 0;
