+++ src/vs/workbench/contrib/chat/browser/chatEditing/notebook/chatEditingNotebookCellEntry.ts
@@ -1,0 +1,345 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { RunOnceScheduler } from '../../../../../../base/common/async.js';
+import { DisposableStore, toDisposable } from '../../../../../../base/common/lifecycle.js';
+import { ITransaction, IObservable, observableValue, autorun, transaction } from '../../../../../../base/common/observable.js';
+import { ObservableDisposable } from '../../../../../../base/common/observableDisposable.js';
+import { themeColorFromId } from '../../../../../../base/common/themables.js';
+import { URI } from '../../../../../../base/common/uri.js';
+import { EditOperation, ISingleEditOperation } from '../../../../../../editor/common/core/editOperation.js';
+import { OffsetEdit } from '../../../../../../editor/common/core/offsetEdit.js';
+import { Range } from '../../../../../../editor/common/core/range.js';
+import { IDocumentDiff, nullDocumentDiff } from '../../../../../../editor/common/diff/documentDiffProvider.js';
+import { DetailedLineRangeMapping } from '../../../../../../editor/common/diff/rangeMapping.js';
+import { TextEdit } from '../../../../../../editor/common/languages.js';
+import { IModelDeltaDecoration, ITextModel, MinimapPosition, OverviewRulerLane } from '../../../../../../editor/common/model.js';
+import { ModelDecorationOptions } from '../../../../../../editor/common/model/textModel.js';
+import { OffsetEdits } from '../../../../../../editor/common/model/textModelOffsetEdit.js';
+import { IEditorWorkerService } from '../../../../../../editor/common/services/editorWorker.js';
+import { IModelContentChangedEvent } from '../../../../../../editor/common/textModelEvents.js';
+import { IConfigurationService } from '../../../../../../platform/configuration/common/configuration.js';
+import { observableConfigValue } from '../../../../../../platform/observable/common/platformObservableUtils.js';
+import { editorSelectionBackground } from '../../../../../../platform/theme/common/colorRegistry.js';
+import { CellEditState } from '../../../../notebook/browser/notebookBrowser.js';
+import { INotebookEditorService } from '../../../../notebook/browser/services/notebookEditorService.js';
+import { NotebookCellTextModel } from '../../../../notebook/common/model/notebookCellTextModel.js';
+import { WorkingSetEntryState } from '../../../common/chatEditingService.js';
+import { IChatResponseModel } from '../../../common/chatModel.js';
+import { pendingRewriteMinimap } from '../chatEditingModifiedFileEntry.js';
+
+
+/**
+ * This is very closely similar to the ChatEditingModifiedDocumentEntry class.
+ * Most of the code has been borrowed from there, as a cell is effectively a document.
+ * Hence most of the same functionality applies.
+ */
+export class ChatEditingNotebookCellEntry extends ObservableDisposable {
+	private static readonly _lastEditDecorationOptions = ModelDecorationOptions.register({
+		isWholeLine: true,
+		description: 'chat-last-edit',
+		className: 'chat-editing-last-edit-line',
+		marginClassName: 'chat-editing-last-edit',
+		overviewRuler: {
+			position: OverviewRulerLane.Full,
+			color: themeColorFromId(editorSelectionBackground)
+		},
+	});
+
+	private static readonly _pendingEditDecorationOptions = ModelDecorationOptions.register({
+		isWholeLine: true,
+		description: 'chat-pending-edit',
+		className: 'chat-editing-pending-edit',
+		minimap: {
+			position: MinimapPosition.Inline,
+			color: themeColorFromId(pendingRewriteMinimap)
+		}
+	});
+
+
+	private _edit: OffsetEdit = OffsetEdit.empty;
+	private _isEditFromUs: boolean = false;
+	public get isEditFromUs(): boolean {
+		return this._isEditFromUs;
+	}
+
+	private _allEditsAreFromUs: boolean = true;
+	public get allEditsAreFromUs(): boolean {
+		return this._allEditsAreFromUs;
+	}
+	private _diffOperation: Promise<any> | undefined;
+	private _diffOperationIds: number = 0;
+
+	private readonly _diffInfo = observableValue<IDocumentDiff>(this, nullDocumentDiff);
+	public get diffInfo(): IObservable<IDocumentDiff> {
+		return this._diffInfo;
+	}
+	private readonly _maxModifiedLineNumber = observableValue<number>(this, 0);
+	readonly maxModifiedLineNumber = this._maxModifiedLineNumber;
+
+	private readonly _editDecorationClear = this._register(new RunOnceScheduler(() => { this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, []); }, 500));
+	private _editDecorations: string[] = [];
+
+	private readonly _diffTrimWhitespace: IObservable<boolean>;
+	protected readonly _stateObs = observableValue<WorkingSetEntryState>(this, WorkingSetEntryState.Modified);
+	readonly state: IObservable<WorkingSetEntryState> = this._stateObs;
+	protected readonly _isCurrentlyBeingModifiedByObs = observableValue<IChatResponseModel | undefined>(this, undefined);
+	readonly isCurrentlyBeingModifiedBy: IObservable<IChatResponseModel | undefined> = this._isCurrentlyBeingModifiedByObs;
+	private readonly initialContent: string;
+
+	constructor(
+		public readonly notebookUri: URI,
+		public readonly cell: NotebookCellTextModel,
+		private readonly modifiedModel: ITextModel,
+		private readonly originalModel: ITextModel,
+		disposables: DisposableStore,
+		@IConfigurationService configService: IConfigurationService,
+		@IEditorWorkerService private readonly _editorWorkerService: IEditorWorkerService,
+		@INotebookEditorService private readonly notebookEditorService: INotebookEditorService
+	) {
+		super();
+		this.initialContent = this.originalModel.getValue();
+		this._register(disposables);
+		this._register(this.modifiedModel.onDidChangeContent(e => {
+			this._mirrorEdits(e);
+		}));
+		this._register(toDisposable(() => {
+			this.clearCurrentEditLineDecoration();
+		}));
+
+		this._diffTrimWhitespace = observableConfigValue('diffEditor.ignoreTrimWhitespace', true, configService);
+		this._register(autorun(r => {
+			this._diffTrimWhitespace.read(r);
+			this._updateDiffInfoSeq();
+		}));
+	}
+
+	public clearCurrentEditLineDecoration() {
+		if (this.modifiedModel.isDisposed()) {
+			return;
+		}
+		this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, []);
+	}
+
+
+	private _mirrorEdits(event: IModelContentChangedEvent) {
+		const edit = OffsetEdits.fromContentChanges(event.changes);
+
+		if (this._isEditFromUs) {
+			const e_sum = this._edit;
+			const e_ai = edit;
+			this._edit = e_sum.compose(e_ai);
+
+		} else {
+
+			//           e_ai
+			//   d0 ---------------> s0
+			//   |                   |
+			//   |                   |
+			//   | e_user_r          | e_user
+			//   |                   |
+			//   |                   |
+			//   v       e_ai_r      v
+			///  d1 ---------------> s1
+			//
+			// d0 - document snapshot
+			// s0 - document
+			// e_ai - ai edits
+			// e_user - user edits
+			//
+			const e_ai = this._edit;
+			const e_user = edit;
+
+			const e_user_r = e_user.tryRebase(e_ai.inverse(this.originalModel.getValue()), true);
+
+			if (e_user_r === undefined) {
+				// user edits overlaps/conflicts with AI edits
+				this._edit = e_ai.compose(e_user);
+			} else {
+				const edits = OffsetEdits.asEditOperations(e_user_r, this.originalModel);
+				this.originalModel.applyEdits(edits);
+				this._edit = e_ai.tryRebase(e_user_r);
+			}
+
+			this._allEditsAreFromUs = false;
+			this._updateDiffInfoSeq();
+
+			const didResetToOriginalContent = this.modifiedModel.getValue() === this.initialContent;
+			const currentState = this._stateObs.get();
+			switch (currentState) {
+				case WorkingSetEntryState.Modified:
+					if (didResetToOriginalContent) {
+						this._stateObs.set(WorkingSetEntryState.Rejected, undefined);
+						break;
+					}
+			}
+
+		}
+	}
+
+	acceptAgentEdits(textEdits: TextEdit[], isLastEdits: boolean, responseModel: IChatResponseModel): void {
+		const notebookEditor = this.notebookEditorService.retrieveExistingWidgetFromURI(this.notebookUri)?.value;
+		if (notebookEditor) {
+			const vm = notebookEditor.getCellByHandle(this.cell.handle);
+			vm?.updateEditState(CellEditState.Editing, 'chatEdit');
+		}
+
+		const ops = textEdits.map(TextEdit.asEditOperation);
+		const undoEdits = this._applyEdits(ops);
+
+		const maxLineNumber = undoEdits.reduce((max, op) => Math.max(max, op.range.startLineNumber), 0);
+
+		const newDecorations: IModelDeltaDecoration[] = [
+			// decorate pending edit (region)
+			{
+				options: ChatEditingNotebookCellEntry._pendingEditDecorationOptions,
+				range: new Range(maxLineNumber + 1, 1, Number.MAX_SAFE_INTEGER, Number.MAX_SAFE_INTEGER)
+			}
+		];
+
+		if (maxLineNumber > 0) {
+			// decorate last edit
+			newDecorations.push({
+				options: ChatEditingNotebookCellEntry._lastEditDecorationOptions,
+				range: new Range(maxLineNumber, 1, maxLineNumber, Number.MAX_SAFE_INTEGER)
+			});
+		}
+
+		this._editDecorations = this.modifiedModel.deltaDecorations(this._editDecorations, newDecorations);
+
+
+		transaction((tx) => {
+			if (!isLastEdits) {
+				this._stateObs.set(WorkingSetEntryState.Modified, tx);
+				this._isCurrentlyBeingModifiedByObs.set(responseModel, tx);
+				this._maxModifiedLineNumber.set(maxLineNumber, tx);
+
+			} else {
+				this._resetEditsState(tx);
+				this._updateDiffInfoSeq();
+				this._maxModifiedLineNumber.set(0, tx);
+				this._editDecorationClear.schedule();
+			}
+		});
+	}
+
+	scheduleEditDecorations() {
+		this._editDecorationClear.schedule();
+	}
+
+	protected _resetEditsState(tx: ITransaction): void {
+		this._isCurrentlyBeingModifiedByObs.set(undefined, tx);
+		this._maxModifiedLineNumber.set(0, tx);
+	}
+
+	public async keep(change: DetailedLineRangeMapping): Promise<boolean> {
+		return this._acceptHunk(change);
+	}
+
+	private async _acceptHunk(change: DetailedLineRangeMapping): Promise<boolean> {
+		this._isEditFromUs = true;
+		try {
+			if (!this._diffInfo.get().changes.includes(change)) {
+				// diffInfo should have model version ids and check them (instead of the caller doing that)
+				return false;
+			}
+			const edits: ISingleEditOperation[] = [];
+			for (const edit of change.innerChanges ?? []) {
+				const newText = this.modifiedModel.getValueInRange(edit.modifiedRange);
+				edits.push(EditOperation.replace(edit.originalRange, newText));
+			}
+			this.originalModel.pushEditOperations(null, edits, _ => null);
+		}
+		finally {
+			this._isEditFromUs = false;
+		}
+		await this._updateDiffInfoSeq();
+		if (this._diffInfo.get().identical) {
+			this._stateObs.set(WorkingSetEntryState.Accepted, undefined);
+		}
+		return true;
+	}
+
+	public async undo(change: DetailedLineRangeMapping): Promise<boolean> {
+		return this._rejectHunk(change);
+	}
+
+	private async _rejectHunk(change: DetailedLineRangeMapping): Promise<boolean> {
+		this._isEditFromUs = true;
+		try {
+			if (!this._diffInfo.get().changes.includes(change)) {
+				return false;
+			}
+			const edits: ISingleEditOperation[] = [];
+			for (const edit of change.innerChanges ?? []) {
+				const newText = this.originalModel.getValueInRange(edit.originalRange);
+				edits.push(EditOperation.replace(edit.modifiedRange, newText));
+			}
+			this.modifiedModel.pushEditOperations(null, edits, _ => null);
+		} finally {
+			this._isEditFromUs = false;
+		}
+		await this._updateDiffInfoSeq();
+		if (this._diffInfo.get().identical) {
+			this._stateObs.set(WorkingSetEntryState.Rejected, undefined);
+		}
+		return true;
+	}
+
+	private _applyEdits(edits: ISingleEditOperation[]) {
+		// make the actual edit
+		this._isEditFromUs = true;
+		try {
+			let result: ISingleEditOperation[] = [];
+			this.modifiedModel.pushEditOperations(null, edits, (undoEdits) => {
+				result = undoEdits;
+				return null;
+			});
+			return result;
+		} finally {
+			this._isEditFromUs = false;
+		}
+	}
+
+	private async _updateDiffInfoSeq() {
+		const myDiffOperationId = ++this._diffOperationIds;
+		await Promise.resolve(this._diffOperation);
+		if (this._diffOperationIds === myDiffOperationId) {
+			const thisDiffOperation = this._updateDiffInfo();
+			this._diffOperation = thisDiffOperation;
+			await thisDiffOperation;
+		}
+	}
+
+	private async _updateDiffInfo(): Promise<void> {
+
+		if (this.originalModel.isDisposed() || this.modifiedModel.isDisposed()) {
+			return;
+		}
+
+		const docVersionNow = this.modifiedModel.getVersionId();
+		const snapshotVersionNow = this.originalModel.getVersionId();
+
+		const ignoreTrimWhitespace = this._diffTrimWhitespace.get();
+
+		const diff = await this._editorWorkerService.computeDiff(
+			this.originalModel.uri,
+			this.modifiedModel.uri,
+			{ ignoreTrimWhitespace, computeMoves: false, maxComputationTimeMs: 3000 },
+			'advanced'
+		);
+
+		if (this.originalModel.isDisposed() || this.modifiedModel.isDisposed()) {
+			return;
+		}
+
+		// only update the diff if the documents didn't change in the meantime
+		if (this.modifiedModel.getVersionId() === docVersionNow && this.originalModel.getVersionId() === snapshotVersionNow) {
+			const diff2 = diff ?? nullDocumentDiff;
+			this._diffInfo.set(diff2, undefined);
+			this._edit = OffsetEdits.fromLineRangeMapping(this.originalModel, this.modifiedModel, diff2.changes);
+		}
+	}
+}
